Moving indirect funding containers to protocols directory
