+++ packages/wallet/src/redux/protocols/indirect-funding/player-b/container.tsx
@@ -1,10 +1,10 @@
 import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
-import ApproveFunding from '../../components/indirect-funding/approve-funding';
-import { FundingStep } from '../../components/indirect-funding/funding-step';
-import * as actions from '../../redux/protocols/indirect-funding/player-a/actions';
-import * as indirectFundingPlayerB from '../../redux/protocols/indirect-funding/player-b/state';
-import { unreachable } from '../../utils/reducer-utils';
+import ApproveFunding from '../../../../components/indirect-funding/approve-funding';
+import { FundingStep } from '../../../../components/indirect-funding/funding-step';
+import * as actions from '../player-a/actions';
+import * as indirectFundingPlayerB from './state';
+import { unreachable } from '../../../../utils/reducer-utils';
 
 interface Props {
   indirectFundingBState: indirectFundingPlayerB.PlayerBState;
